Product assignment story added

+++ app/src/main/java/io/ramani/ramaniWarehouse/data/stockassignment/model/ProductCategory.kt
@@ -4,5 +4,5 @@ data class ProductCategory(
     val _id: String,
     val categoryId: String,
     val name: String,
-    val unitPrice: Int
+    val unitPrice: Double
 )
